+++ dali/operators/sequence/sequence_rearrange_test.cc
@@ -1,4 +1,4 @@
-// Copyright (c) 2019-2021, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// Copyright (c) 2019-2022, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -119,7 +119,7 @@ void SeqRearrangeVerify(const testing::TensorListWrapper& input,
   auto in = input.CopyTo<CPUBackend>();
   auto out = output.CopyTo<CPUBackend>();
   auto order = args.at(testing::ArgumentKey("new_order")).GetValue<std::vector<int>>();
-  for (size_t i = 0; i < out->num_samples(); i++) {
+  for (int i = 0; i < out->num_samples(); i++) {
     CheckRearrange(out->tensor<int>(i), in->tensor_shape(i), out->tensor_shape(i), order);
   }
 }
